Implement --shared option when adding a target
When an executable target is marked as shared, all child targets with a dynamicLinkPrefix (not documented yet) will have their symbols exported. When a library target is marked shared, the same thing happens, and the target is built as a DLL (or shared library).

This flag is needed when adding the plytool target to a new build folder on Windows. Otherwise, the plytool that gets built won't be able to load its instantiator DLLs, and the repo registry won't be initialized.