Fix unqualified '// ply extern' names
Fox example, in Instantiator.inl you can now write
    // ply extern sdl2.homebrew
instead of
    // ply extern <reponame>.sdl2.homebrew
The second form has the advantage of being unambiguous when the same extern name is defined in multiple repos. It would be a good idea to issue a warning when ambiguity is detected (currently no warning is issued) but that will involve deciding which cases are considered ambiguous and which cases are not.