Implement plytool --auto option
'plytool build' and 'plytool run' now both accept --auto as a command line option. When this option is used, and there are no build folders that contain the specified target, PlyTool will automatically create a new build folder and add the target before building. If there's a unique build folder that already contains the specified target, that build folder will be used instead.