Fix TextConverter::convert()
"Hello\n😋🍺🍕\n" was missing the trailing linefeed when converting from UTF-8 to UTF-16. This new implementation is better. It now takes advantage of DecodeResult::Status::Truncated (recently added) to decide when data should be buffered internally. It also consumes as much of srcBuf as possible, even if flush is false. The only difference when flush is true is that if the input stream ends with a truncated multibyte character, it will be decoded as a series of single code units instead of waiting for more input.