BuildTarget is no longer a subclass of Dependency
Dependency now has an optional BuildTarget pointer. If this pointer is non-null, the Dependency will add a build target to the generated build system. Currently, it is null for all externs and non-null for all instantiated modules. The intention is to let it be non-null for some externs in the future as well.