Fix #11 by exempting src/math/tests from codegen
This is a quick fix. It will be easier to prevent such bugs in the future by (1) implementing automated tests and (2) making 'plytool codegen' opt-in instead of opt-out. (Longer term, the 'plytool codegen' command should cope with preprocessor macros in a more appropriate way; it currently supports macro expansion, but skips all #include and #define directives. )